Restructure settings pages and move actual pages to the rest of pages under src/pages/daos/[daoAddress]
