+++ src/components/pages/SafeSettings/ERC20TokenContainer.tsx
@@ -1,10 +1,10 @@
 import { Box, Flex, Text } from '@chakra-ui/react';
 import { useTranslation } from 'react-i18next';
-import { useFractal } from '../../../../providers/App/AppProvider';
-import { AzoriusGovernance } from '../../../../types';
-import { formatCoin } from '../../../../utils';
-import { DisplayAddress } from '../../../ui/links/DisplayAddress';
-import { BarLoader } from '../../../ui/loaders/BarLoader';
+import { useFractal } from '../../../providers/App/AppProvider';
+import { AzoriusGovernance } from '../../../types';
+import { formatCoin } from '../../../utils';
+import { DisplayAddress } from '../../ui/links/DisplayAddress';
+import { BarLoader } from '../../ui/loaders/BarLoader';
 import { SettingsSection } from './SettingsSection';
 
 export function ERC20TokenContainer() {
@@ -15,11 +15,7 @@ export function ERC20TokenContainer() {
   const { votesToken } = azoriusGovernance;
 
   return (
-    <SettingsSection
-      title={t('governanceTokenTitle')}
-      descriptionHeader={t('governanceTokenTitle')}
-      descriptionContent={t('governanceTokenDescription')}
-    >
+    <SettingsSection title={t('governanceTokenTitle')}>
       {votesToken ? (
         <Flex
           justifyContent="space-between"
