+++ src/components/ui/modals/ModalProvider.tsx
@@ -4,8 +4,8 @@ import { useTranslation } from 'react-i18next';
 import { Address } from 'viem';
 import { UnsavedChangesWarningContent } from '../../../pages/daos/[daoAddress]/roles/edit/unsavedChangesWarningContent';
 import { ProposalTemplate } from '../../../types';
-import AddSignerModal from '../../pages/DaoSettings/components/Signers/modals/AddSignerModal';
-import RemoveSignerModal from '../../pages/DaoSettings/components/Signers/modals/RemoveSignerModal';
+import AddSignerModal from '../../pages/SafeSettings/Signers/modals/AddSignerModal';
+import RemoveSignerModal from '../../pages/SafeSettings/Signers/modals/RemoveSignerModal';
 import DraggableDrawer from '../containers/DraggableDrawer';
 import { DAOSearch } from '../menus/DAOSearch';
 import { ConfirmModifyGovernanceModal } from './ConfirmModifyGovernanceModal';
