+++ src/i18n/locales/en/settings.json
@@ -1,11 +1,8 @@
 {
   "settings": "Settings",
+  "settingsPageTitle": "{{daoName}} Settings",
   "governanceTokenTitle": "Governance Token",
   "governanceERC721TokenTitle": "Governance NFT Token(s)",
-  "governanceParametersTitle": "Governance Voting Parameters",
-  "governanceParametersDescription": "Quorum, Voting Period, Timelock Period, Execution Period - these parameters controlling sensible proposal timings and are impacting community participation and speed at which your DAO can run proposals.",
-  "governanceTokenDescription": "A governance token is an ERC-20 token that grants holders the right to vote and participate in an organization's decision-making process.",
-  "governanceERC721TokenDescription": "A governance NFT is an ERC-721 token that grants holders the right to vote and participate in an organization's decision-making process.",
   "governanceTokenNameLabel": "Name",
   "governanceTokenSymbolLabel": "Symbol",
   "governanceTokenWeightLabel": "Token Weight",
@@ -17,14 +14,10 @@
   "guardTitle": "Guard",
   "modulesAndGuardsTitle": "Modules and Guards",
   "permissionsTitle": "Permissions",
-  "modulesAndGuardsDescription1": "Modules are added contracts that extend a Safe's capabilities.",
-  "modulesAndGuardsDescription2": "A Freeze Guard is added to all child Safes created on Decent. A Freeze Guard allows the parent Safe to block the child Safe's access to funds and freeze active proposals.",
   "daoSettingsGeneral": "General",
   "daoSettingsGovernance": "Governance",
   "daoMetadataName": "Safe Name",
   "daoMetadataSnapshot": "Snapshot Space",
-  "daoMetadataDescriptionTitle": "Safe Metadata & offchain",
-  "daoMetadataDescriptionText": "Activities or transactions that occur outside of the blockchain network or aren't stored onchain.",
   "proposeChanges": "Propose Change",
   "labelMultisig": "Multisig",
   "labelLinearErc20": "ERC-20",
